# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v1.3.1
servers:
  - url: http://localhost:8080
    description: local
  - url: /histudy
    description: dev
security:
  - bearerAuth: [ ]
paths:
...
  /api/auth/login:
    get:
      security: [ ]
      tags:
        - 인증 API
      summary: 로그인
      operationId: login
      parameters:
        - name: sub
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserLogin'
  /api/auth/token:
    post:
      security: [ ]
      tags:
        - 인증 API
      summary: 토큰 재발급
      operationId: issueAccessToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenForm'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TokenInfo'

  # ===================
  # User Management APIs
  # ===================
...
  /api/forms:
    post:
      tags:
        - 스터디 신청 API
      summary: 스터디 신청
      operationId: applyForStudy_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyForm'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApplyFormDto'
      deprecated: true
  /api/v2/forms:
    post:
      tags:
        - 스터디 신청 API
      summary: 스터디 신청
      operationId: applyForStudy
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyFormV2'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApplyFormDto'

  # ===================
  # Course Management APIs
  # ===================
  /api/courses:
    get:
      tags:
        - 강의 관리 API
      summary: 강의 목록 조회
      operationId: getCourses
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseDto'
    post:
      tags:
        - 강의 관리 API
      summary: 강의 목록 업로드
      operationId: importCourses
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: OK
  /api/courses/delete:
    post:
      tags:
        - 강의 관리 API
      summary: 강의 삭제
      operationId: deleteCourse
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseIdDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int32

  # ===================
  # Study Group/Team APIs
  # ===================
...
  /api/admin/edit-user:
    post:
      tags:
        - 관리자 API
      summary: 유저 정보 수정
      operationId: editUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserEdit'
        required: true
      responses:
        '200':
          description: OK
  /api/admin/form:
    delete:
      tags:
        - 관리자 API
      summary: 특정 유저 지원폼 삭제
      operationId: deleteForm
      parameters:
        - name: sid
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
...
  /api/admin/groupReport/{id}:
    get:
      tags:
        - 관리자 API
      summary: 특정 그룹 보고서 조회
      operationId: getTeamReports
      parameters:
        - name: id
          in: path
          description: 그룹 아이디
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TeamReportDto'
  /api/admin/group:
    delete:
      tags:
        - 관리자 API
      summary: 그룹 삭제
      operationId: deleteTeam
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamIdDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int32
      deprecated: true
components:
  schemas:
    ApplyFormV2:
      type: object
      properties:
        friendIds:
          type: array
          description: List of friend added to apply form
          example: [ 1, 2 ]
          items:
            type: integer
            description: List of friend added to apply form
            format: int64
        courseIds:
          type: array
          description: List of course added to apply form
          example: [ 1, 2 ]
          items:
            type: integer
            description: List of course added to apply form
            format: int64
    ApplyFormDto:
      type: object
      properties:
        friends:
          type: array
          description: List of friend added to apply form
          items:
            $ref: '#/components/schemas/UserBasic'
        courses:
          type: array
          description: List of course added to apply form
          items:
            $ref: '#/components/schemas/CourseInfo'
    CourseInfo:
      type: object
      properties:
        id:
          type: number
          description: Course ID
          example: 1
        name:
          type: string
          description: Course Name
          example: Software Engineering
        prof:
          type: string
          description: Course Professor
          example: Prof. John Doe
        code:
          type: string
          description: Course Code
          example: CSEE 4111
        year:
          type: number
          description: Course Year
          example: 2021
        semester:
          type: number
          description: Course Semester
          example: 1
      description: List of course added to apply form
    UserBasic:
      type: object
      properties:
        id:
          type: number
          description: User ID
          example: 1
        name:
          type: string
          description: User name
          example: John Doe
        sid:
          type: string
          description: User student ID
          example: '211234567'
      description: List of friend added to apply form
...
    JwtPair:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      description: Token pairs
    UserLogin:
      type: object
      properties:
        isRegistered:
          type: boolean
          description: Registration status
          example: true
        tokenType:
          type: string
          description: Token type
          example: 'Bearer '
        tokens:
          $ref: '#/components/schemas/JwtPair'
        role:
          type: string
          description: Role
          example: USER
...
    ApplyForm:
      type: object
      properties:
        friendIds:
          type: array
          description: List of friend added to apply form
          example: [ '21800111', '21900111' ]
          items:
            type: string
            description: List of friend added to apply form
            example: '"21800111","21900111"'
        courseIds:
          type: array
          description: List of course added to apply form
          example: [ 1, 2 ]
          items:
            type: integer
            description: List of course added to apply form
            format: int64
    CourseIdDto:
      type: object
      properties:
        id:
          type: number
          description: Course ID
          example: 1
    TokenForm:
      type: object
      properties:
        grantType:
          type: string
          description: Token Type
          example: refresh_token
        refreshToken:
          type: string
          description: Token
          example: <sample-jwt-token>
    TokenInfo:
      type: object
      properties:
        tokenType:
          type: string
        grantType:
          type: string
          enum:
            - ACCESS_TOKEN
            - REFRESH_TOKEN
        token:
          type: string
    UserEdit:
      type: object
      properties:
        id:
          type: integer
          description: User ID
          format: int64
        team:
          type: integer
          description: Group tag
          format: int32
          example: 112
        name:
          type: string
          description: User name
          example: John Doe
        sid:
          type: string
          description: User student ID
          example: '21800012'
...
    CourseDto:
      type: object
      properties:
        courses:
          type: array
          description: List of courses
          items:
            $ref: '#/components/schemas/CourseInfo'
...
    ReportBasic:
      type: object
      properties:
        id:
          type: number
          description: Report ID
          example: 1
        title:
          type: string
          description: Report Title
          example: Week 15 Report
        regDate:
          type: string
          description: Report Last Modified Date
          example: '2021-06-01 00:00:00'
        totalMinutes:
          type: number
          description: Total minutes of the report
          example: 60
        thumbnail:
          type: string
          description: Thumbnail of the report
          example: https://histudy.s3.ap-northeast-2.amazonaws.com/2021-06-01-00-00-00-1
      description: Reports information
      example: 120
    TeamReportDto:
      type: object
      properties:
        group:
          type: number
          description: Team ID
          example: 1
          deprecated: true
        tag:
          type: integer
          description: Team tag number
          format: int32
          example: 1
        members:
          type: array
          description: Team members
          items:
            $ref: '#/components/schemas/UserBasic'
        totalTime:
          type: number
          description: Total time studied
          example: 300
        reports:
          type: array
          description: Reports information
          example: 120
          items:
            $ref: '#/components/schemas/ReportBasic'
    TeamIdDto:
      type: object
      properties:
        groupId:
          type: number
          description: Team ID
          example: 1
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT