# src/main/java/edu/handong/csee/histudy/controller/form/TokenForm.java
package edu.handong.csee.histudy.controller.form;

import lombok.*;

@Getter
@AllArgsConstructor
@NoArgsConstructor(access = AccessLevel.PRIVATE)
public class TokenForm {

    private String grantType;

    private String refreshToken;
}