# src/main/java/edu/handong/csee/histudy/dto/TeamIdDto.java
package edu.handong.csee.histudy.dto;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class TeamIdDto {

    private long groupId;
}